Display generated tags separately
