+++ frontend/src/components/VMediaInfo/VMediaDetails.vue
@@ -1,5 +1,5 @@
 <template>
-  <section class="flex flex-col gap-y-6">
+  <section class="flex flex-col gap-y-6 md:gap-y-8">
     <header class="flex flex-row items-center justify-between">
       <h2 class="heading-6 md:heading-5">
         {{ $t(`mediaDetails.${media.frontendMediaType}Info`) }}
@@ -9,10 +9,18 @@
     <div class="flex flex-col items-start gap-6 md:flex-row">
       <slot name="thumbnail" />
 
-      <div class="flex w-full flex-grow flex-col gap-6">
-        <p v-if="media.description">{{ media.description }}</p>
-        <VMediaTags :tags="media.tags" :media-type="media.frontendMediaType" />
-        <VMetadata v-if="metadata" :metadata="metadata" />
+      <div class="flex flex-col gap-6 md:gap-8">
+        <div
+          class="flex w-full flex-grow flex-col items-start gap-6 md:flex-row"
+        >
+          <p v-if="media.description">{{ media.description }}</p>
+          <VMetadata v-if="metadata" :metadata="metadata" />
+        </div>
+        <VMediaTags
+          :tags="media.tags"
+          :media-type="media.frontendMediaType"
+          :provider="media.provider"
+        />
       </div>
     </div>
   </section>
