+++ frontend/src/locales/scripts/en.json5
@@ -820,7 +820,13 @@
     imageInfo: "Image information",
     audioInfo: "Audio track information",
     tags: {
-      title: "Tags",
+      generated: {
+        heading: "Generated tags",
+        tooltip: "Tags generated by an AI model",
+      },
+      source: {
+        heading: "Source tags",
+      },
       showMore: "Show more",
       showLess: "Show less",
     },
